Controlpanels Plone4 support
