+++ src/plone/restapi/controlpanels/types.py
@@ -14,14 +14,16 @@ from plone.restapi.controlpanels.interfaces import IDexterityTypesControlpanel
 import plone.protect.interfaces
 from zope.interface import noLongerProvides
 from plone.restapi.interfaces import IPloneRestapiLayer
+from Products.CMFPlone.utils import getFSVersionTuple
 
+PLONE5 = getFSVersionTuple()[0] >= 5
 
 @adapter(Interface, IPloneRestapiLayer)
 @implementer(IDexterityTypesControlpanel)
 class DexterityTypesControlpanel(RegistryConfigletPanel):
     schema = Interface
     configlet_id = "dexterity-types"
-    configlet_category_id = "plone-content"
+    configlet_category_id = "plone-content" if PLONE5 else 'Plone'
 
     def add(self, names):
         data = json_body(self.request)
