+++ src/plone/restapi/deserializer/configure.zcml
@@ -44,8 +44,6 @@
     <adapter factory=".atfields.ReferenceFieldDeserializer" />
   </configure>
 
-  <configure zcml:condition="have plone-5">
-    <include package=".controlpanels" />
-  </configure>
+  <include package=".controlpanels" />
 
 </configure>
