+++ src/plone/restapi/services/configure.zcml
@@ -7,13 +7,10 @@
 
   <include package=".auth" />
   <include package=".actions" />
-  <include package=".addons"
-      zcml:condition="have plone-5"/>
+  <include package=".addons" />
   <include package=".breadcrumbs"/>
   <include package=".content"/>
-  <configure zcml:condition="have plone-5">
-    <include package=".controlpanels" />
-  </configure>
+  <include package=".controlpanels" />
   <include package=".copymove"/>
   <include package=".database"/>
   <include package=".discussion"/>
